Fix error of import (ndinterp)
